Avoid race between updating stepper and ViewPager instantiating the steps.
It was assuming the view pager has layed out all the steps at the time that it calls back to its listeners,
but that is not always true (Handler.post() is happening before ViewPager has inserted the step fragments into the
fragment manager).

This solution waits for the ViewPager to complete its layout before updating StepperLayout.